skills/deepagents-setup-configuration/assets/templates/deep-agent-simple/pyproject.toml
[project]
name = "deep-agent-simple"
version = "0.1.0"
description = "Simple Deep Agent example"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.0",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
